requirements.txt
streamlit
openai
scikit-learn
pandas
numpy
python-docx
PyPDF2
python-dotenv
